backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-allauth==65.12.0
djangorestframework_simplejwt==5.5.1
django-filter==25.2

gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1

# One line is enough for psycopg:
psycopg[binary,pool]==3.2.10

requests==2.32.5
cryptography==46.0.2

# ML Dependencies
numpy==2.3.3
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.2
keras==3.11.3
tensorflow==2.20.0